Better handling invalid subcommands ##shell
