+++ libr/core/cmd_help.inc.c
@@ -662,14 +662,17 @@ R_API void r_core_clippy(RCore *core, const char *msg) {
 
 #include "visual_riu.inc.c"
 
+const char iuhelp[] =
+"Usage: ?iu fieldname(type,command,value)\n"
+"  Types: string, button, title, run\n"
+"Examples:\n"
+"'?iu name(string,?i;yp,test) addr(string,f~...) ok(button) cancel(button)\n"
+"'?iu addr(string,f~...) hexdump(run,x 32@k:riu.addr) ok(button)\n"
+"Values for every field are saved in the global SdbKv database (see `k` command)\n";
+
 static int cmd_qiu(RCore *core, const char *input) {
 	if (!*input || *input == '?') {
-		r_cons_printf ("Usage: ?iu fieldname(type,command,value)\n");
-		r_cons_printf ("  Types: string, button, title, run\n");
-		r_cons_printf ("Examples:\n");
-		r_cons_printf ("'?iu name(string,?i;yp,test) addr(string,f~...) ok(button) cancel(button)\n");
-		r_cons_printf ("'?iu addr(string,f~...) hexdump(run,x 32@k:riu.addr) ok(button)\n");
-		r_cons_printf ("Values for every field are saved in the global SdbKv database (see `k` command)\n");
+		r_cons_print (iuhelp);
 		return 0;
 	}
 	RIU *riu = riu_new (core, input);
@@ -1092,6 +1095,7 @@ static int cmd_help(void *data, const char *input) {
 			break;
 		default:
 			r_cons_printf ("0x%"PFMT64x"\n", n);
+			break;
 		}
 		r_core_return_value (core, n); // redundant
 		break;
@@ -1256,6 +1260,9 @@ static int cmd_help(void *data, const char *input) {
 		case '2':
 			r_cons_printf ("%d\n", R2_VERSION_PATCH);
 			break;
+		default:
+			r_core_return_invalid_command (core, "?V", input[1]);
+			break;
 		}
 		break;
 	case 'l': // "?l"
@@ -1263,18 +1270,18 @@ static int cmd_help(void *data, const char *input) {
 			for (input += 2; input[0] == ' '; input++);
 			r_core_return_value (core, strlen (input));
 		} else {
-			for (input++; input[0] == ' '; input++);
+			input = r_str_trim_head_ro (input + 1);
 			r_core_return_value (core, strlen (input));
 			r_cons_printf ("%" PFMT64d "\n", core->num->value);
 		}
 		break;
 	case 'X': // "?X"
-		for (input++; input[0] == ' '; input++);
+		input = r_str_trim_head_ro (input + 1);
 		n = r_num_math (core->num, input);
 		r_cons_printf ("%"PFMT64x"\n", n);
 		break;
 	case 'x': // "?x"
-		for (input++; input[0] == ' '; input++);
+		input = r_str_trim_head_ro (input + 1);
 		if (*input == '-') {
 			ut8 *out = malloc (strlen (input) + 1);
 			if (out) {
@@ -1295,9 +1302,7 @@ static int cmd_help(void *data, const char *input) {
 			}
 			r_cons_newline ();
 		} else {
-			if (*input == ' ') {
-				input++;
-			}
+			input = r_str_trim_head_ro (input);
 			for (i = 0; input[i]; i++) {
 				r_cons_printf ("%02x", input[i]);
 			}
@@ -1434,7 +1439,7 @@ static int cmd_help(void *data, const char *input) {
 				  free (d);
 			}
 			break;
-		case 'p':
+		case 'p': // "?ep"
 			  {
 				  char *word, *str = strdup (r_str_trim_head_ro (input + 2));
 				  char *legend = strchr (str, ',');
@@ -1461,7 +1466,7 @@ static int cmd_help(void *data, const char *input) {
 					  text[j] = word;
 					  j++;
 				  }
-				  int size = r_config_get_i (core->config, "hex.cols");
+				  const int size = r_config_get_i (core->config, "hex.cols");
 				  r_print_pie (core->print, r_list_length (list), nums, (const char**)text, size);
 				  free (text);
 				  r_list_free (list);
@@ -1477,40 +1482,45 @@ static int cmd_help(void *data, const char *input) {
 			free (newmsg);
 			}
 			break;
-		case 0:
+		case 0: // "?e"
 			r_cons_newline ();
 			break;
-		default:
+		case '?': // "?e?"
 			r_core_cmd_help (core, help_msg_question_e);
 			break;
+		default:
+			r_core_return_invalid_command (core, "?e", input[1]);
+			break;
 		}
 		break;
-	case 's': { // "?s" sequence from to step
-		ut64 from, to, step;
-		char *p, *p2;
-		for (input++; *input == ' '; input++);
-		p = strchr (input, ' ');
-		if (p) {
-			*p = '\0';
-			from = r_num_math (core->num, input);
-			p2 = strchr (p+1, ' ');
-			if (p2) {
-				*p2 = '\0';
-				step = r_num_math (core->num, p2 + 1);
-			} else {
-				step = 1;
+	case 's': // "?s" sequence from to step
+		{
+			input = r_str_trim_head_ro (input + 1);
+			char *p = strchr (input, ' ');
+			if (p) {
+				*p = '\0';
+				ut64 from = r_num_math (core->num, input);
+				char *p2 = strchr (p+1, ' ');
+				int step = 0;
+				if (p2) {
+					*p2 = '\0';
+					step = r_num_math (core->num, p2 + 1);
+				}
+				if (step < 1) {
+					step = 1;
+				}
+				ut64 to = r_num_math (core->num, p + 1);
+				for (; from <= to; from += step) {
+					r_cons_printf ("%"PFMT64d" ", from);
+				}
+				r_cons_newline ();
 			}
-			to = r_num_math (core->num, p + 1);
-			for (;from <= to; from += step)
-				r_cons_printf ("%"PFMT64d" ", from);
-			r_cons_newline ();
 		}
 		break;
-	}
 	case 'P': // "?P"
 		if (core->io->va) {
 			ut64 o, n = (input[0] && input[1])?
-				r_num_math (core->num, input+2): core->offset;
+				r_num_math (core->num, input + 2): core->offset;
 			RIOMap *map = r_io_map_get_paddr (core->io, n);
 			if (map) {
 				o = n + r_io_map_begin (map) - map->delta;
@@ -1571,7 +1581,7 @@ static int cmd_help(void *data, const char *input) {
 				{
 				char foo[1024];
 				r_cons_flush ();
-				for (input+=2; *input == ' '; input++);
+				input = r_str_trim_head_ro (input + 2);
 				// TODO: r_cons_input()
 				snprintf (foo, sizeof (foo) - 1, "%s: ", input);
 				r_line_set_prompt (foo);
@@ -1585,21 +1595,21 @@ static int cmd_help(void *data, const char *input) {
 				 r_cons_any_key (NULL);
 				 break;
 			case 'y': // "?iy"
-				 for (input += 2; *input == ' '; input++);
+				 input = r_str_trim_head_ro (input + 2);
 				 r_core_return_value (core, r_cons_yesno (1, "%s? (Y/n)", input));
 				 break;
 			case 'u':
 				 r_core_return_value (core, cmd_qiu (core, r_str_trim_head_ro (input + 2)));
 				 break;
 			case 'n': // "?in"
-				 for (input += 2; *input == ' '; input++);
+				input = r_str_trim_head_ro (input + 2);
 				 r_core_return_value (core, r_cons_yesno (0, "%s? (y/N)", input));
 				 break;
 			default: {
 				char foo[1024];
 				r_cons_flush ();
-				for (input++; *input == ' '; input++);
-				// TODO: r_cons_input()
+				input = r_str_trim_head_ro (input + 1);
+				// TODO: use r_cons_input()
 				snprintf (foo, sizeof (foo) - 1, "%s: ", input);
 				r_line_set_prompt (foo);
 				r_cons_fgets (foo, sizeof (foo), 0, NULL);
@@ -1650,10 +1660,12 @@ static int cmd_help(void *data, const char *input) {
 		}
 		break;
 	case '\0': // "?"
-	default:
 		// TODO #7967 help refactor
 		r_core_cmd_help (core, help_msg_intro);
 		r_core_cmd_help (core, help_msg_root);
+		break;
+	default:
+		r_core_return_invalid_command (core, "?", input[0]);
 		break;
 	}
 	return 0;
