+++ libr/core/yank.c
@@ -1,31 +1,19 @@
-/* radare - LGPL - Copyright 2009-2022 - pancake & dso */
+/* radare - LGPL - Copyright 2009-2024 - pancake & dso */
 
 #include <r_core.h>
 
-/*
- * perform_mapped_file_yank will map in a file and yank from offset the number of len bytes from
- * filename.  if the len is -1, the all the bytes are mapped into the yank buffer.
- */
-static bool perform_mapped_file_yank(RCore *core, ut64 offset, ut64 len, const char *filename);
-static ut32 find_next_char(const char *input, char b);
-static ut32 consume_chars(const char *input, char b);
-
 static ut32 find_next_char(const char *input, char b) {
 	ut32 i = 0;
-	if (input) {
-		for (; *input != b; i++, input++) {
-			/* nothing */
-		}
+	for (; *input != b; i++, input++) {
+		/* nothing */
 	}
 	return i;
 }
 
 static ut32 consume_chars(const char *input, char b) {
 	ut32 i = 0;
-	if (input) {
-		for (; *input == b; i++, input++) {
-			/* nothing */
-		}
+	for (; *input == b; i++, input++) {
+		/* nothing */
 	}
 	return i;
 }
@@ -121,7 +109,7 @@ R_API bool r_core_yank_set_str(RCore *core, ut64 addr, const char *str, ut32 len
 R_API bool r_core_yank(RCore *core, ut64 addr, int len) {
 	ut64 curseek = core->offset;
 	if (len < 0) {
-		R_LOG_ERROR ("cannot yank negative bytes");
+		R_LOG_ERROR ("Cannot yank negative bytes");
 		return false;
 	}
 	if (len == 0) {
@@ -146,16 +134,15 @@ R_API bool r_core_yank(RCore *core, ut64 addr, int len) {
 /* Copy a zero terminated string to the clipboard. Clamp to maxlen or blocksize. */
 R_API bool r_core_yank_string(RCore *core, ut64 addr, int maxlen) {
 	ut64 curseek = core->offset;
-	ut8 *buf = NULL;
 	if (maxlen < 0) {
-		R_LOG_ERROR ("cannot yank negative bytes");
+		R_LOG_ERROR ("Cannot yank negative bytes");
 		return false;
 	}
 	if (addr != core->offset) {
 		r_core_seek (core, addr, true);
 	}
 	/* Ensure space and safe termination for largest possible string allowed */
-	buf = calloc (1, core->blocksize + 1);
+	ut8 *buf = calloc (1, core->blocksize + 1);
 	if (!buf) {
 		return false;
 	}
@@ -196,115 +183,102 @@ R_API bool r_core_yank_paste(RCore *core, ut64 addr, int len) {
 R_API bool r_core_yank_to(RCore *core, const char *_arg) {
 	ut64 len = 0;
 	ut64 pos = -1;
-	char *str, *arg;
 	bool res = false;
-
-	while (*_arg == ' ') {
-		_arg++;
-	}
-	arg = strdup (_arg);
-	str = strchr (arg, ' ');
+	char *arg = strdup (r_str_trim_head_ro (_arg));
+	char *str = strchr (arg, ' ');
 	if (str) {
 		str[0] = '\0';
 		len = r_num_math (core->num, arg);
 		pos = r_num_math (core->num, str + 1);
 		str[0] = ' ';
 	}
-	if (!str || pos == -1 || len == 0) {
-		free (arg);
-		return res;
-	}
-	if (r_core_yank (core, core->offset, len) == true) {
-		res = r_core_yank_paste (core, pos, len);
+	if (str && pos != -1 && len > 0) {
+		if (r_core_yank (core, core->offset, len) == true) {
+			res = r_core_yank_paste (core, pos, len);
+		}
 	}
 	free (arg);
 	return res;
 }
 
 R_API bool r_core_yank_dump(RCore *core, ut64 pos, int format) {
-	bool res = false;
 	int i = 0;
 	int ybl = core->yank_buf ? r_buf_size (core->yank_buf): 0;
-	if (ybl > 0) {
-		if (pos < ybl) {
-			switch (format) {
-			case '8':
-				for (i = pos; i < r_buf_size (core->yank_buf); i++) {
-					r_cons_printf ("%02x", r_buf_read8_at (core->yank_buf, i));
-				}
-				r_cons_newline ();
-				break;
-			case 'j': {
-				PJ *pj = r_core_pj_new (core);
-				if (!pj) {
-					break;
-				}
-				pj_o (pj);
-				pj_kn (pj, "addr", core->yank_addr);
-				RStrBuf *buf = r_strbuf_new ("");
-				for (i = pos; i < r_buf_size (core->yank_buf); i++) {
-					r_strbuf_appendf (buf, "%02x", r_buf_read8_at (core->yank_buf, i));
-				}
-				pj_ks (pj, "bytes", r_strbuf_get (buf));
-				r_strbuf_free (buf);
-				pj_end (pj);
-				r_cons_println (pj_string (pj));
-				pj_free (pj);
-				break;
-			}
-			case '*':
-				//r_cons_printf ("yfx ");
-				r_cons_printf ("wx ");
-				for (i = pos; i < r_buf_size (core->yank_buf); i++) {
-					r_cons_printf ("%02x", r_buf_read8_at (core->yank_buf, i));
-				}
-				//r_cons_printf (" @ 0x%08"PFMT64x, core->yank_addr);
-				r_cons_newline ();
-				break;
-			default:
-				r_cons_printf ("0x%08" PFMT64x " %" PFMT64d " ",
-						core->yank_addr + pos,
-						r_buf_size (core->yank_buf) - pos);
-				for (i = pos; i < r_buf_size (core->yank_buf); i++) {
-					r_cons_printf ("%02x", r_buf_read8_at (core->yank_buf, i));
-				}
-				r_cons_newline ();
-			}
-			res = true;
-		} else {
-			R_LOG_ERROR ("Position exceeds buffer length");
-		}
-	} else {
+	if (ybl < 1) {
 		if (format == 'j') {
 			r_cons_printf ("{}\n");
 		} else {
 			R_LOG_ERROR ("No buffer yanked yet");
 		}
+		return false;
 	}
-	return res;
+	if (pos >= ybl) {
+		R_LOG_ERROR ("Position exceeds buffer length");
+		return false;
+	}
+	switch (format) {
+	case '8':
+		for (i = pos; i < r_buf_size (core->yank_buf); i++) {
+			r_cons_printf ("%02x", r_buf_read8_at (core->yank_buf, i));
+		}
+		r_cons_newline ();
+		break;
+	case 'j': {
+		PJ *pj = r_core_pj_new (core);
+		if (!pj) {
+			break;
+		}
+		pj_o (pj);
+		pj_kn (pj, "addr", core->yank_addr);
+		RStrBuf *buf = r_strbuf_new ("");
+		for (i = pos; i < r_buf_size (core->yank_buf); i++) {
+			r_strbuf_appendf (buf, "%02x", r_buf_read8_at (core->yank_buf, i));
+		}
+		pj_ks (pj, "bytes", r_strbuf_get (buf));
+		r_strbuf_free (buf);
+		pj_end (pj);
+		r_cons_println (pj_string (pj));
+		pj_free (pj);
+		break;
+	}
+	case '*':
+		r_cons_printf ("'wx ");
+		for (i = pos; i < r_buf_size (core->yank_buf); i++) {
+			r_cons_printf ("%02x", r_buf_read8_at (core->yank_buf, i));
+		}
+		r_cons_newline ();
+		break;
+	default:
+		r_cons_printf ("0x%08" PFMT64x " %" PFMT64d " ",
+				core->yank_addr + pos,
+				r_buf_size (core->yank_buf) - pos);
+		for (i = pos; i < r_buf_size (core->yank_buf); i++) {
+			r_cons_printf ("%02x", r_buf_read8_at (core->yank_buf, i));
+		}
+		r_cons_newline ();
+	}
+	return true;
 }
 
 R_API bool r_core_yank_hexdump(RCore *core, ut64 pos) {
-	bool res = false;
-	int ybl = r_buf_size (core->yank_buf);
+	const int ybl = r_buf_size (core->yank_buf);
 	if (ybl > 0) {
 		if (pos < ybl) {
 			ut8 *buf = R_NEWS (ut8, ybl - pos);
-			if (!buf) {
-				return false;
+			if (buf) {
+				r_buf_read_at (core->yank_buf, pos, buf, ybl - pos);
+				r_print_hexdump (core->print, pos,
+					buf, ybl - pos, 16, 1, 1);
+				free (buf);
+				return true;
 			}
-			r_buf_read_at (core->yank_buf, pos, buf, ybl - pos);
-			r_print_hexdump (core->print, pos,
-				buf, ybl - pos, 16, 1, 1);
-			res = true;
-			free (buf);
 		} else {
 			R_LOG_ERROR ("Position exceeds buffer length");
 		}
 	} else {
 		R_LOG_ERROR ("No buffer yanked yet");
 	}
-	return res;
+	return false;
 }
 
 R_API bool r_core_yank_cat(RCore *core, ut64 pos) {
@@ -313,14 +287,13 @@ R_API bool r_core_yank_cat(RCore *core, ut64 pos) {
 		if (pos < ybl) {
 			ut64 sz = ybl - pos;
 			char *buf = R_NEWS (char, sz);
-			if (!buf) {
-				return false;
+			if (buf) {
+				r_buf_read_at (core->yank_buf, pos, (ut8 *)buf, sz);
+				r_cons_write (buf, sz);
+				r_cons_newline ();
+				free (buf);
+				return true;
 			}
-			r_buf_read_at (core->yank_buf, pos, (ut8 *)buf, sz);
-			r_cons_write (buf, sz);
-			r_cons_newline ();
-			free (buf);
-			return true;
 		}
 		R_LOG_ERROR ("Position exceeds buffer length");
 	} else {
@@ -330,6 +303,7 @@ R_API bool r_core_yank_cat(RCore *core, ut64 pos) {
 }
 
 R_API bool r_core_yank_cat_string(RCore *core, ut64 pos) {
+	R_RETURN_VAL_IF_FAIL (core, false);
 	int ybl = r_buf_size (core->yank_buf);
 	if (ybl > 0) {
 		if (pos < ybl) {
@@ -345,7 +319,7 @@ R_API bool r_core_yank_cat_string(RCore *core, ut64 pos) {
 			free (buf);
 			return true;
 		}
-		R_LOG_ERROR ("Position exceeds buffer length");
+	//	R_LOG_ERROR ("Position exceeds buffer length");
 	} else {
 		r_cons_newline ();
 	}
@@ -353,11 +327,11 @@ R_API bool r_core_yank_cat_string(RCore *core, ut64 pos) {
 }
 
 R_API bool r_core_yank_hud_file(RCore *core, const char *input) {
+	R_RETURN_VAL_IF_FAIL (core, false);
 	if (R_STR_ISEMPTY (input)) {
 		return false;
 	}
-	input = r_str_trim_head_ro (input + 1);
-	char *buf = r_cons_hud_file (input);
+	char *buf = r_cons_hud_file (r_str_trim_head_ro (input + 1));
 	ut32 len = buf? strlen ((const char *) buf) + 1: 0;
 	bool res = r_core_yank_set_str (core, R_CORE_FOREIGN_ADDR, buf, len);
 	free (buf);
@@ -365,26 +339,23 @@ R_API bool r_core_yank_hud_file(RCore *core, const char *input) {
 }
 
 R_API bool r_core_yank_hud_path(RCore *core, const char *input, int dir) {
-	char *buf = NULL;
-	ut32 len = 0;
-	int res;
-	for (input++; *input == ' '; input++) {
-		/* nothing */
-	}
-	buf = r_cons_hud_path (input, dir);
-	len = buf? strlen ((const char *) buf) + 1: 0;
-	res = r_core_yank_set_str (core, R_CORE_FOREIGN_ADDR, buf, len);
+	R_RETURN_VAL_IF_FAIL (core, false);
+	char *buf = r_cons_hud_path (r_str_trim_head_ro (input), dir);
+	ut32 len = buf? strlen ((const char *) buf) + 1: 0;
+	int res = r_core_yank_set_str (core, R_CORE_FOREIGN_ADDR, buf, len);
 	free (buf);
 	return res;
 }
 
 R_API void r_core_yank_unset(RCore *core) {
+	R_RETURN_IF_FAIL (core);
 	r_buf_free (core->yank_buf);
 	core->yank_buf = NULL;
 	core->yank_addr = UT64_MAX;
 }
 
 R_API bool r_core_yank_hexpair(RCore *core, const char *input) {
+	R_RETURN_VAL_IF_FAIL (core, false);
 	if (R_STR_ISEMPTY (input)) {
 		return false;
 	}
@@ -424,7 +395,7 @@ R_API bool r_core_yank_file_ex(RCore *core, const char *input) {
 
 	ut64 next = find_next_char (inp + adv, ' ');
 	if (next) {
-		inp[adv+next] = 0;
+		inp[adv + next] = 0;
 	} else {
 		R_LOG_ERROR ("File must be specified");
 		free (inp);
